requirements.txt
django>5.1,<5.3
djangorestframework>3.15,<3.16
djangorestframework-simplejwt>5.2,<5.3
pytest-django>4.11.0,<4.11.9
pillow>11.3.0,<11.3.9